+++ utils/anthem_build_support/anthem_build_support/products/catch.py
@@ -20,6 +20,14 @@ from .. import diagnostics, shell
 
 class Catch(product.Product):
     def do_build(self):
+        # Check whether the headers of the current version are already
+        # installed.
+        if os.path.exists(os.path.join(self.workspace.install_root,
+                                       'include',
+                                       'catch.hpp')) \
+                and os.path.exists(self.build_dir):
+            return
+
         # Delete the old build directory of Catch.
         shell.rmtree(self.build_dir)
